added authentication to deleteReview

+++ src/controllers/reviewController.js
@@ -196,6 +196,39 @@ exports.updateReview = async (req, res) => {
 exports.deleteReview = async (req, res) => {
   const reviewId = req.params.reviewId;
 
+  //Get username from token
+  let token;
+  // Grab the Authorization header
+  const authHeader = req.headers.authorization;
+
+  // Check it contains JWT token and extract the token
+  if (authHeader && authHeader.startsWith("Bearer")) {
+    token = authHeader.split(" ")[1];
+  }
+
+  // Get userId from token
+  const payload = jwt.verify(token, process.env.JWT_SECRET);
+  const userId = payload.userId;
+  const role = payload.role;
+  console.log(payload);
+
+  //Check if correct user or ADMIN
+
+  //get review from DB
+  const [reviewMatch] = await sequelize.query(
+    `SELECT fk_user_id FROM review WHERE Id = $reviewId;`,
+    {
+      bind: {
+        reviewId: reviewId,
+      },
+      type: QueryTypes.SELECT,
+    }
+  );
+
+  if (reviewMatch.fk_user_id !== userId || role !== userRoles.ADMIN) {
+    throw new UnauthorizedError("You can only delete your own reviews");
+  }
+
   const [results, metadata] = await sequelize.query(
     `DELETE FROM review WHERE id = $reviewId RETURNING *`,
     {
